Fix a couple of typos.
    compiler/main/config.cpp
In the assert, comparison (==) rather than assignment (=) was obviously intended.  

The assertion succeded up to now because the map returns the expression that was
just inserted.  However, that makes the sense of the assertion, "the map
contains {something} at the location of the key where I just inserted a
key/value pair."  With the typo corrected, the test is, "the map contains {the
value I just inserted} at the location of the key where I just inserted a
key/value pair."

    compiler/passes/codegen.cpp
*


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22300 3a8e244f-b0f2-452b-bcba-4c88e055c3ca