Fix the massive memory leak in fileIO.chpl (primers) [reviewed by vass]
This reduces the memory leaked in fileIO.chpl from 1040194752 bytes to 192
bytes.  The remaining leaks are due to leaking local strings, which is a known
problem.

The central change is the addition of FLAG_INSERT_AUTO_DESTROY to the assignment
case in insertRetMove() in normalize.cpp:360.  This causes the internal
temporary receiving the return value through assignment to be autodestroyed
before the containing functions returns.

By UMM rules, if the value of a variable of record type is established through
assignment, then that variable "owns" the copy it receives.  In contrast, if the
value of the variable is established through a "move" primitive (as it is in the
second case following in insertRetMove()), then ownership is shared between the
two copies.  In that case, it suffices to destroy just the original and not the
copy.  Destroying both can lead to double-deletion or premature frees.  

To preserve the behavior of internally reference-counted types, the added
auto-destroy flags have to be removed from return values of those types.  The
clause added ca. line 90 in callDestructors.cpp takes care of this removal.  To
get a successful regression test run, I found that I only needed to mention
arrays and domains (not distributions) as types requiring the exceptional
behavior.  Further memory leaks testing may reveal that we need to add the
distribution flag as well.

The remaining changes were small tweaks required to get a clean test run,
additional comments, and a typo correction.


DETAILS

    compiler/AST/iterator.cpp
Explicit removal of the DefExpr for a disused return value variable (symbol)
caused an assertion failure.  This resulted directly from the main change in
this patch, because we are now reusing this symbol in the call to the
autoDestroy function, where these are called for by the newly-inserted
FLAG_INSERT_AUTO_DESTROY.

That the return statement referencing it is being removed is no longer a
sufficient condition for the variable symbol being completely unused in the body
of a given function.  Therefore, it is premature to remove that variable's
DefExpr.  The normal post-pass cleanup will find and remove the DefExpr if it is
truly unused at that point in the translation.

    compiler/passes/normalize.cpp
When the value of the return value variable is established through assignment,
the return value variable owns that copy.  The copy must therefore be destroyed
before the function returns.

The way record-return-by-value is currently implemented, an autoCopy is added in
the caller, to copy the returned value into the local scope.  The transformation
changeRetToArgAndClone() pulls this autoCopy call inside the called routine, so
the copy happens before the called function returns and the return value
variable goes out of scope.   For a reference-counted type, the autoCopy bumps
the reference count, and this happens before the return value variable is
destroyed.  Therefore, the reference count of a reference-counted type will be
correct on return, as follows:
  retval = x;                         // Ref count in retval == 1.
  (*retArg) = chpl__autoCopy(retval); // Ref count in retval == 2.
  chpl__autoDestroy(retval);          // Ref count in retval == 1.
  return;                             // retval goes out of scope
                                      // but (*retArg) contains a bitwise
                                      // copy of retval that persists.

    compiler/resolution/callDestructors.cpp
Added a comment about why we have to remove FLAG_INSERT_AUTO_DESTROY from the
return value of initCopy functions in particular.

Added special case code for when the return value is of array or domain type.
This sandboxes the above changes so they do not affect how memory management is
done for these internally reference-counted types.  It may also be necessary to
remove FLAG_INSERT_AUTO_DESTROY from distribution types, but early testing did
not show this to be the case.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22303 3a8e244f-b0f2-452b-bcba-4c88e055c3ca