Update .bad files due to r22292.
Application of pragma "compiler generated" to the default implementation of
chpl__initCopy() (ChapelBase.chpl:1018) and its generic expansions to disappear
from the backtrace of calls noted in a resolution failure.  Resolution failures
that were formerly attributed to the first line of the record that causes the
generic expansion of that default chpl__initCopy() are now attributed to the
source line that calls said chpl__initCopy().



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22306 3a8e244f-b0f2-452b-bcba-4c88e055c3ca