If include/tasks/*/tasks-*.h declares func(), include/chpl-tasks.h should not.
[Reviewed by Lydia.]

The change in r22294 broke building with the PGI compiler, because the
static inline definition for chpl_task_getSubloc() in tasks-fifo.h,
along with an undecorated prototype declaration in chpl-tasks.h, caused
the compiler to produce a global function definition in every .o file.
This led to multiply-defined symbols at link time.

To fix this, here we add a standardized technique for an implementation
to have its tasks-*.h file replace chpl-tasks.h as the defining entity
for any of the tasking interface functions.  If the implementation's
tasks-*.h #defines CHPL_TASK_func_IMPL_DECL, then chpl-tasks.h assumes
that the implementation has declared func(), and does not do so itself.
For now this is only supported for
  chpl_task_getSubloc()
  chpl_task_setSubloc()
  chpl_task_getRequestedSubloc()
but we can add more as needed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22313 3a8e244f-b0f2-452b-bcba-4c88e055c3ca