Fix incompatible pointer type warnings on Darwin
[Reviewed by Micheal]

This fixes the incompatible pointer type warnings within qio when size_t and
ssize_t are defined as being longs, while [u]int64_t are long longs.

The short term solution implemented in this patch is to replace any pointers to
size_t and ssize_t with pointers to [u]int64_t. The true fix for this problem
would be to preserve the typedefs through to the generated code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22315 3a8e244f-b0f2-452b-bcba-4c88e055c3ca