"export" implies "local args". [reviewed by gbt]
This is an optional part of the "wide export" fix.

  Motivation:

chpl_memTracking_returnConfigVals() (a function exported from MemTracking.chpl)
was using wide pointers to write values back into narrow pointers.  valgrind
complains about this, because half of the wide pointer overwrites memory not
owned by the actual argument.

  Rationale:

In a discussion with Brad, we concluded that the formal arguments of an
"export"ed function should always be narrow, because external code is not
currently expected to support a concept of wide pointer.

  Implementation:

In the parser, FLAG_LOCAL_ARGS is now applied to all functions marked as
"export" functions.  Internally-generated functions that are marked for export
are similarly marked with the "local args" flag.

To ensure that the new invariant "export" => "local args" is maintained in the
AST, I added a new check.  At the same time, I added a check for "extern" =>
"local args".

  Details:

: modules/internal/MemTracking.chpl
"local args" is now implied by "export".

: compiler/include/flags_list.h
Added more verbose comments to FLAG_LOCAL_ARGS.

: compiler/AST/build.cpp
: compiler/parser/chapel.ypp
: compiler/passes/addInitGuards.cpp
: compiler/passes/buildDefaultFunctions.cpp
FLAG_LOCAL_ARGS is added to every function marked for export.

: compiler/main/checks.cpp
Added checks to ensure:
  "extern" => "local args"
  "export" => "local args"

[tested on test/ (std)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22320 3a8e244f-b0f2-452b-bcba-4c88e055c3ca