Fix the wide export error. [reviewed by ]
This is the essential part of the "wide export" fix.

  Motivation:

chpl_memTracking_returnConfigVals() (a function exported from MemTracking.chpl)
was using wide pointers to write values back into narrow pointers.  valgrind
complains about this, because half of the wide pointer overwrites memory not
owned by the actual argument.

  Rationale:

In a discussion with Brad, we concluded that the formal arguments of an
"export"ed function should always be narrow, because external code is not
currently expected to support a concept of wide pointer.

I first tried narrowing the args to exported functions directly, by removing
those args from the set stored in wideRefMap in widenRefs().  However, this gave
argument type mismatches in the generated C code, because the existing
assignments in the routine of interest (chpl_memTracking_returnConfigVals())
expect wide refs on the LHS.  It appeared that the narrowWideRefs optimization
would allow these LHS symbols to be narrow and propagate that to the exported
interface, provided that the left side of an assignment can, in general, be
narrowed.

Assignments from wide to narrow and narrow to wide are supported, so only the
wideness of the symbol itself (the LHS) matters.  The wideness of the RHS can be
safely ignored.  This is similar to what is done for the "move" and "op=" (and
other) primitives.  So I duplicated that case for assignment.

  Implementation:

The top loop in narrowSym() scans functions which overwrite the value of a given
symbol.  Certain functions are weeded out as not forcing the return value to be
wide.  I added assignment to this list.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22330 3a8e244f-b0f2-452b-bcba-4c88e055c3ca