Fix the inner-extern.good file. [Reviewed by Vass]
Similar to r22324, this test's .good file didn't match what the program prints
and as a result of r22314, the value of c_sublocid_any has been changed from -1
to -2.  The future will still fail in the configuration where this was noticed,
because it is printing a warning now that the error message has been resolved.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22340 3a8e244f-b0f2-452b-bcba-4c88e055c3ca