Adjusted test/extern/vass/PREDIFF to handle the PrgEnv-gnu backend compiler.
[trivial]

At the moment c_sublocid_any.inner-extern.chpl fails because
in the generated code we have something like:

  T = &c_sublocid_any;

where T is not declared with 'const', whereas c_sublocid_any is; then
(some of) the backend compilers complain about discarding the 'const'
qualifier.

In most configurations we test this is an error (due to -Werror),
which was recognized by the PREDIFF, which ensured a clean match against .bad.
Under PrgEnv-gnu this is currently just a warning.

This change makes PREDIFF handle warnings just like it handles errors,
so that we still have a clean match against .bad.

PREDIFF still prints "got error(s)" instead of "got error(s) or warning(s)"
because I did not want the trouble of modifying all .bad files.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22341 3a8e244f-b0f2-452b-bcba-4c88e055c3ca