Added isSubtype(), isProperSubtype() to standard modules and the spec. Also added a test.
[reviewed by Tom]

Synopsis:
  isSubtype(type sub, type super) param
  isProperSubtype(type sub, type super) param
return whether 'sub' and 'super' are in a (proper) subtype relationship.

As explained in the addition to the spec, two types are in a subtype
relationship when either they both are the same or they both are
class (or record) types with 'sub' inheriting from 'super',
perhaps indirectly.

The output of the test contains some line-number-less warnings.
Those are artifacts of the present compiler's warning reporting.
When that changes, the .good will need an adjustment.

While creating test cases, I noticed that we do not report an error
for a union inheriting from a record. The spec does not allow that
and the compiled program crashed. I'll leave that for another day.

While there also - in the parser TUNION misses "@$.comment = yylloc.comment; 
yylloc.comment = NULL;" which are present for TCLASS and TRECORD.
Something we might want to fix.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22342 3a8e244f-b0f2-452b-bcba-4c88e055c3ca