Fixed an optimization think-o in _array.reindex().
[reviewed by Tom]

reindex() has had the following optimization since r20252:
If the array's domain is identical to the domain that reindex needs to
reindex to, just return the original array (with ref counting increment).

The "identical" check is simply pointer comparison. That check is skipped
when the types of the array's domain and the reindex-to domain are different.

The condition for skipping compared the type of the reindex-to domain against
the type of the array's domain descriptor, resulting in always skipping.

This commit makes it compare the types of the two domain descriptors instead.
The comparison is generalized to the case where the static type of one
of the descriptor is a subtype of that of the other descriptor.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22343 3a8e244f-b0f2-452b-bcba-4c88e055c3ca