Factor the buildForallLoopExpr() function. [reviewed by Lydia]
The buildForallLoopExpr() function creates a serial iterator, a leader iterator
and a follower iterator.  Those actions were factored out as separate functions.

The order of the arguments in each function signature was adjusted to match the
order of the arguments in corresponding calls.  This saves us from generating an
_order_wrap_ wrapper each time we instantiate one of these iterator functions,
and saves us having to execute the order wrapper.  (Most likely this only has a
performance impact when inlining is turned off.)

[tested with test/ (std), primers (no-local)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22352 3a8e244f-b0f2-452b-bcba-4c88e055c3ca