Add a log/linear toggle button to each graph. [not reviewed]
Adds a button in the footter below each graph, which toggles the vertical scale
between linear and logarithmic.  

I moved the gspacer so it is now a footer rather than a header.  The new button
appears inside the footer.  (There is more room there for other options.) 
I grouped HTML elements creation, the setting of their attributes and insertion
into the DOM -- for readability.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22354 3a8e244f-b0f2-452b-bcba-4c88e055c3ca