Skip reprivatizeIterators() when --local.
[reviewed by Tom]

Privatization is counter-productive when there is only a single locale,
so skip it in this one case.

Background: reprivatizeIterators() transforms those fields of iterator
classes that hold privatized classes. It changes those fields to hold
the pids of those classes, and augments the accesses to those fields
with conversions pid <-> class pointer.

A more general privatization elimination for --local might be
worthwhile, but is outside the scope of this patch.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22356 3a8e244f-b0f2-452b-bcba-4c88e055c3ca