Fix the valgrind regression in refcnt.chpl following r22303 and r22312. [reviewed by Lydia]
The fix for the massive memory leak in fileIO.chpl (r22303) improved the
behavior of users/ferguson/refcnt.chpl, but caused it (and a few other tests) to
hang unpredictably.  valgrind revealed memory allocation errors, which were due
to a mismatch in the wideness of the formal and actual arguments passed to
chpl_memTracking_returnConfigVals(), and this mismatch was repaired by r22320.

refcnt.chpl was still failing in the scope labelled "copy initialization 2".
This was fixed by enabling the copyInit function (r22331).  Following this,
refcnt.chpl still had valgrind errors due to our failure to call autoCopy on a
(user-defined) record passed by value to a begin block.  This patch fixes that
problem.

  The Problem:

A user-defined record type that implements reference counting will reclaim the
memory in the reference-counted object if the count goes to zero.  The following
code snippet shows the reference counts in a code snippet where autoCopy is not
called on a record passed by value to a begin block.
  0 {
  1   var r:R;
  1   begin {
  1     test_one(r, false);
  1   }
  0 }
Since r goes out of scope at the end of the enclosing block, the reference count
will go to zero and its contents will be reclaimed.  If all this happens before
the spawned task (to run test_one()) commences, then the data reference in the
copy of r passed to test_one() will be a dangling reference.  At that point, the
behavior of the program is unpredictable, since that freed memory can be
overwritten arbitrarily.

  The Solution:

Clearly, it is necessary to increment the reference count in r prior to spawning
the new task, and code must be added at the end of the task itself, to decrement
the reference count so that memory can be reclaimed.  There is already code
present in bundleArgs (parallel.cpp) to perform these insertions for internally
reference-counted types.  The code was expanded to include user-defined record
types passed by value.

  The Bug Tail:

I made several tweaks to the code before pinning down the error in bundleArgs.
In refcnt.chpl, the reference count was not going to zero as we exited each
scope.  This was cured by adding the "insert auto destroy" flag to each formal
temp when it is created [functionResolution.cpp:2965 (revised)].

After this change, the ref count for sync variables was going below zero, so I
had to add FLAG_SYNC to the list of flags on return value types that have the
"insert auto destroy" flag culled from then in cullAutoDestroyFlags()
[callDestructors.cpp:39 (revised)].

Array reference counts were still going negative.  In order to clarify this
error, I factored and slightly re-wrote the buildForLoopExpr() utility.  For no
apparent reason, the order of the arguments in calls to the three iterator
functions did not match.  In each, both the call and the function are
compiler-generated, so we can make them match.  By comparison with other
iterator function signatures, I decided that the order of arguments in the call
are consistent with other iterator signatures, so I chose to update the function
signatures for the compiler-generated iterator functions instead.  This change
is contained in a companion patch [r ].

Next I made the main change in this patch, adding autoCopy and autoDestroy calls
when records are passed by value to a begin function.  This resulted in 21
regressions on a standard test run.  The first of these was solved by adding a
missing "auto destroy fn" pragma on the corresponding function in
types/records/sungeun/classInRecord.chpl.

Two regressions were resolved by separating the code for user-defined records
from that for internally reference-counted types in bundleArgs [parallel.cpp:223
(revised)].  The clause for internally reference-counted types adds the autoCopy
only once but adds the autodestroy for each spawned task.

Other tests were failing due to increased memory leaks.  A tweak to my patch
fixed this.  Early deletion failures were prevented by removing the "insert auto
destroy" flag from temporaries of tuple type, but this turned out to be the
wrong solution.

In the test domainInRecordInClass, an early-deletion error was caused by the
fact that we were copying the value of a formal_tmp into a record field (in a
default constructor) without calling the autoCopy function.  But the formal_tmp
was still being deleted.  This makes the reference count 1 less than it should
be and leads to an early-deletion failure.

This and the remaining failing cases were resolved by adding a clause to
cullAutoDestroyFlags() [callDestructors.cpp:43 (revised)] to find formal temps
of (non-ref) record type used to initialize the fields of a record or class
inside a default constructor.  From those formal temporaries, the "insert auto
destroy" flag is removed.  This restores the autoCopy/autoDestroy equilibrium
within those compiler-generated (default) constructors.


DETAILS

  test/classes/figueroa/RecordDestructor1b.good
Since we are now calling autoCopy and autoDestroy on user-defined records passed
to begin blocks, the number of times we call the destructor for R increased by 1.

  test/types/records/sungeun/classInRecord.chpl
The user autoDestroy function defined for myR must be marked "auto destroy fn".

  test/users/ferguson/refcnt.good
Enabling the chpl__initCopy function changed reference counts throughout the run.

  compiler/passes/scopeResolve.cpp
Added a comment on how autoCopies and autoDestroys are kept in balance by
cullAutoDestroyFlags().

  compiler/passes/parallel.cpp
Narrowed the interface to bundleArgs.  The function being called can be
extracted from fcall.
Added comments.
Added a clause to insert autoCopy and autoDestroy calls on record arguments
passed by value to a begin block.

  compiler/resolution/functionResolution.cpp
Moved the clause that adds the "const" flag to certain formal temps closer to
where it is used in addLocalCopiesAndWritebacks().
Added the "insert auto destroy" flag on all formal temps that have blank, const
or const in intent and are not sync types and are not internally ref-counted
types.
In the "else" clause of the case for blank and const intent, the "insert auto
destroy" flag is removed if the formal is created without calling
chpl__autoCopy().

  compiler/resolution/callDestructors.cpp
FLAG_SYNC was added to the cases of return types for which autoDestroy calls are
not inserted.
A clause was added to look for places in a default constructor function where
record fields are initialized from formal temps without calling chpl__autoCopy.
The "insert auto destroy" flag is removed from these formal temps, so the
autoCopies and autoDestroys balance.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22359 3a8e244f-b0f2-452b-bcba-4c88e055c3ca