Add the ability to run multiple trials for performance tests
This patch provides the ability to run multiple trials for performance tests.
This can be utilized by adding a new file, PERFTRIALOPTS or
{execname}.perftrialopts, that contains the number of trials as its only
contents. To actually get any difference in the performance graphs, a few new
options in the .graph file are needed:

    generate: [avg,med], ...
    displayrange: true

An example has been included for meteor

The generate option matches up with the corresponding perfkey or file options
and merges multiple results for any day using the given strategy.
(avg=average, med=median)

The displayrange option enables the customBars setting in dygraph, this causes
a shaded area to surround the line, going from min time to max time per trial
(it makes the graph look like the NY vs SF Temp. graph in Dygraph's examples).

Details
=======
The changes to sub_test merely make it run the performance test multiple times,
appending the results to the dat file. This means that the dat file format does
not change, other than that multiple trials appear as lines with the same date.
By doing this, old dat files are entirely compatible with the new script.

The changes to genGraphs perform the work of merging multiple entries of the
same date. If there are multiple entries, they are reduced by one of three
strategies:
    avg: Finds the average of the trials
    med: Finds the median of the trials
    default: Selects the first result of the trials
This is perhaps an overcautious default, but it is the (somewhat) expected
behavior before this patch.

When using displayrange = true, the CSV format that Dygraph expects changes
slightly.  The excepted form becomes:
    "date,low1;val1;high1,low2;val2;high2,..."
Where low and high will be the boundaries of the shaded area, while the line
travels through val.

If displayrange = false, the old CSV format is used. In practice this shouldn't
matter as the CSV files are currently regenerated with each performance run.
(Allowing us to switch over to the new format quite easily)

Further Thoughts
================
Something to look at in the future would be changing the default value on these
options to have multiple trials, always show ranges, and always calculate the
average. But, for now I have left everything functionally equivalent to the
prior version.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22361 3a8e244f-b0f2-452b-bcba-4c88e055c3ca