Fix issues with genGraphs
[Reviewed by Vass]

There were a couple of issues the changes to genGraphs in r22361. The primary
problem was forgetting that some of the perfkeys are not numbers. This is
solved by adding a function called try_parse_float and using that instead of
directly casting the values.

In addition, this patch fixes an issue that caused incorrect CSV files to be
generated when multiple perfkeys were pulled from a single file. To solve this,
the mergedLines list was broken out into each DatFileClass instance.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22364 3a8e244f-b0f2-452b-bcba-4c88e055c3ca