Fix typo in genGraphs
[Not reviewed]

genGraphs is currently performing a check to ensure the dat file had data for a
given date by checking

fields[df.mykeys[i]] != '-'

But, as each element of fields is a list, this would never be true. Checking
with

'-' not in fields[df.mykeys[i]]

gives the expected behavior. This only became a problem when running with all
of the historical performance data.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22365 3a8e244f-b0f2-452b-bcba-4c88e055c3ca