Cleanup to the disambiguation by match process.
This patch breaks the disambiguation process into several functions, cleans up
the interfaces of the functions involved, encapsulates some of the data into
objects, re-names variables for clarity, and updates the tracing information to
make it more helfpul.  In addition, new tests were added that trigger most of
the clauses in the disambiguation process.

Reviewed by Tom Hildebrandt.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22376 3a8e244f-b0f2-452b-bcba-4c88e055c3ca