Revert jacobi-no-inline to a regular test.
[Reviewed conversationally in the deep dive meeting this morning.]

This undoes r22309 and r22274, which turned jacobi-no-inline into a
future and limited it to being tested with comm=gasnet, respectively.
It turned out that the problem causing it to fail was not the already
known issue with coerce/promote wrappers, but rather the wide export
issue fixed by Tom's r22330.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22377 3a8e244f-b0f2-452b-bcba-4c88e055c3ca