Make some tests leak-free, by deleting allocated objects. [reviewed by vass]
A number of test cases from the spec were leaking a small number of bytes due to
the tests failing to clean up after themselves.  I did a general sweep of the
spec for class variables that were initialized with a 'new' and not cleaned up
with a corresponding 'delete'.  Then I added the missing deletes.

In most cases, I added the delete statement in a chapelnoprint or chapelpost
environment, so it is not printed in the spec.  It might otherwise detract from
the main point of the example.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22383 3a8e244f-b0f2-452b-bcba-4c88e055c3ca