Add accelerator locale model
[reviewed by greg]

This adds an accelerator locale model to support targeting accelerators.
Primarily this adds the correct backend compiler flags to enable accelerator
code compilation (currently for OpenACC, but this is meant to be general to
other methods for targeting accelerators.)  

This adds the environment variable CHPL_ACC_CODEGEN. The variable only really
matters when CHPL_LOCALE_MODEL=acc, otherwise it is set to none and ignored.
Everything that goes along with adding an environment variable is also included
in this commit such as the driver changes, module changes, and the environment
variable scripts. The idea behind this variable is that over time we may have
other methods of targeting accelerators. Currently we're working to support
OpenACC, but OpenMP, CUDA, OpenCL, etc may be used and as such we need some way
to specify the code generation method that is actually going to be used. Most
of the changes are a result of adding a new environment variable and the rest
are related to adding the actual model (and the makefile changes to throw the
backend flags for openacc.) 

For now this just copies the flat locale model from modules and runtime. 

The search path for the locale module was also updated to search the
subdirectories of the acc directory. 

This also includes doc updates to chplenv since CHPL_ACC_CODEGEN was added, and
to localeModels since a new locale model was added. Note that the docs changes
represent the current state of the accelerator locale model and accelerator
support and they will certainly have to be changed as that support is updated.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22384 3a8e244f-b0f2-452b-bcba-4c88e055c3ca