Removing accidental commit
I left some code in driver.cpp when I was looking into propagating changes 
to the chapel config params to the equivalent variables in driver.cpp. The code
set comm, segment, and substrate to none in the driver which basically forced
local compilation always, which explains why this only affected tests with
comm set. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22389 3a8e244f-b0f2-452b-bcba-4c88e055c3ca