Disable gets/puts for single locale when using numa locale model
[reviewed by David]

This commit replaces codegen of chpl_gen_comm_get/put() with straight
memcpy() in cases where we use wide pointers for single locale
compilation (and do not request --no-local).  Currently, this happens
only when using the numa locale model.  This improves numa locale
model performance by ~2x for the stream and stream-ep benchmarks.

I also factored out the direct comparison of CHPL_LOCALE_MODEL into a
function called forceWidePtrs().  The function is called by
requireWideReferences() and requireOutlinedOn(), as well as in a new
function called forceWidePtrsForLocal() which is used in
codegenAssign() as described above.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22391 3a8e244f-b0f2-452b-bcba-4c88e055c3ca