Bringing darwin testing more up-to-date with reality
Retired the darwin failures that Kyle fixed last week.  Also ran through
other regressions that were listed to see if they were still failing
and for the same reasons.  Note that we haven't done a full run on darwin
for quite some time now, so this may just be a subset of all failures.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22397 3a8e244f-b0f2-452b-bcba-4c88e055c3ca