Enable the reclamation of sync variable array elements [reviewed by Lydia]
In dsiDestroyData() [DefaultRectangular.chpl:480], the param function
chpl__maybeAutoDestroyed() queries the type of an array element to see if it
should be reclaimed through a call to chpl__autoDestroy().

Right now, the query returns "false" for sync and single variables.  Since sync
and single variables are implemented as class instances, this means that they
will be leaked.

Since dsiDestroyData is called only when the containing array is being destroyed
(its ref count having gone to zero), it seems reasonable that this reclamation
can be done without creating any dangling references (at least, not within the
implementation of the array type itself).  

This patch reverses the sense of that test, so that sync and single array
elements will be reclaimed when the containing array is destroyed.

[tested on test/ with (std)]
[tested on domains/ distributions/ multilocale/ performance/ with CHPL_COMM=gasnet]
[tested on examples/ with -valgrind]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22403 3a8e244f-b0f2-452b-bcba-4c88e055c3ca