Updated Classes.tex with "delete nil is OK". [reviewed by vass]
Also, updated the declaration example to show that two class variable names can
be aliases of one another.  Due to that change, I could move the recently-added
"delete c;" statement into the chapelpost section, so it does not detract from
the example.

I added to the delete example, showing that deleting a nil variable has no
effect. I added the warning that an object (name) should not be used after it
has been deleted, and buttressed this with examples.

Note that this example works because the class it uses does not define a
nontrivial destructor.  The compiler must still be modified to meet the spec in
that case.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22404 3a8e244f-b0f2-452b-bcba-4c88e055c3ca