Add options to start_test to control number of trials and related settings
[Reviewed by Lydia]

This patch enables the changes from r22361 to the performance tests by default.
It also adds flags to start_test to control the options on a global scale.

The new defaults and options for performance tests are:

Number of trials: 1
-num-trials <number>

Display trial range: True
-no-display-graph-range (boolean flag)

Generation strategy for multiple trials: avg
-graphs-gen-default <avg|min|max|med>

util/cron/nightly now has a -num-trials setting as well
The nightly cron scripts that run performance tests have this set to 3

There are also a couple of minor fixes in this patch:

sub_test now looks for PERFNUMTRIALS rather than PERFTRIALOPTS.

Display settings were changed in perfgraph.js to show decimal points if
the graphs are of large enough scale to be in scientific notation.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22419 3a8e244f-b0f2-452b-bcba-4c88e055c3ca