Support for release-over-release performance testing
[reviewed by Elliot ages ago, but never committed (!)]

This commit adds a new script:

     util/pastPerformance/testReleasesPerformance

which can be used to do a performance testing sweep of the current
test suite (or, more generally, an arbitrary version of the test
suite) across past releases.

This differs from what we do in normal nightly performance testing
since that captures the performance on each night for that night's
version of the test.  Thus, this script has the advantage that it
factors out improvements to the benchmark code itself, permitting
focus to be placed on changes in the Chapel implementation's
performance.

Specific changes to existing scripts to support this are:

* Made start_test support the setting of an optional env var
  CHPL_TEST_UTIL_DIR to indicate what util/ directory should be
  used for testing scripts.  This permits us to use the current
  test system and scripts with an older release.

* Made sub_test use CHPL_TEST_UTIL_DIR to determine where the
  utility scripts live

* Made sub_test look at (optional) CHPL_TEST_PERF_DATE to
  determine what date to pass through to computePerfStats

* Made computePerfStats support an optional new argument that
  indicates the date to use in generating a .dat file entry.
  This is used to generate performance data points in the past
  rather than using today's date.

* Fixed a misleading path in start_test that reflects a directory
  reorganization that took place years ago.


Some future TODO/cleanup opportunities:

o unify specifying dates of release branches with the performance
  graphing system to avoid duplication

o separate past release performance from current HEAD in .dat files
  and/or support splicing/dicing of .dat files to be able to
  append latest performance onto past without re-running all of
  the past ones.

o automatically name the default perfdat directory based on the 
  current revision being tested.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22426 3a8e244f-b0f2-452b-bcba-4c88e055c3ca