Extend replication of global constants to other types
[reviewed by vass]

This commit fixes a bug/oversight in which we failed to replicate global
constants of simple types across locales as advertised in the spec. It adds
enums, imaginaries, and complexes to the list of things to be replicated.
It also changes the test used for booleans from a simple comparison to
dtBool (which I think would only cover default-width boolean types?) to use
is_bool_type(), more symmetrically with other types.

One case that I was curious about but did not investigate (due to lack of
time, difficulty in verifying, and perhaps good motivating examples) was
whether global const class variables are replicated (which would correspond
to replicating the reference to the object, not the object itself).  So that
remains an open question for future investigation.

Added a test to lock this behavior in.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22429 3a8e244f-b0f2-452b-bcba-4c88e055c3ca