Add a nightly.standard script to run the testing suite with the defaults and add a script for tests on mbp-kbrady
[Reviewed by Sung]

This patch adds a nightly.standard script, along the lines of nightly.fast and
nightly.numa, to run the testing suite with the default options. This script is
meant to be called from a machine.<host> file, which is described further in
the README and machine.template.

A pair of files, machine.mbp-kbrady.* have been added for normal testing runs
on darwin. MWF runs with clang, while TR runs with gcc.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22430 3a8e244f-b0f2-452b-bcba-4c88e055c3ca