Fix read-after-free on domains used in nonblocking on calls. [reviewed by vass]
Because they run asynchronously, non-blocking "on" function calls should behave
like "begin" function calls with respect to reference counting.  That is, the
reference count on internally reference-counted instances should be incremented
prior to the call.  Those reference counts should be decremented in the tasking
function before _downEndCount is called.

The fix is to add the case of non-blocking "on" tasking functions to the
conditional that performs this reference count manipulation for "begin" blocks
in bundleArgs() [parallel.cpp:191].

The new case causes an autodestroy call to be added for the dummy_locale_arg
argument that is passed to an on_fn(), and that caused the AST to be malformed
following execution of passArgsToNestedFns().  The solution was to remove the
"break" statement in that function on line 1108 of parallel.cpp.  This has the
effect of removing (only) the offending autoDestroy call.

The second test case failed because the autoCopy being carefully added above was
getting removed by the removeUnnecessaryAutoCopyCalls optimization.  So, I added
the FLAG_NECESSARY_AUTO_COPY to prevent the removal of necessary autoCopy calls
such as the ones being added in bundleArgs.

With these changes, the failing test case provided by Vass passes.


DETAILS

    compiler/include/flags_list.h
Added the new FLAG_NECESSARY_AUTO_COPY flag, to mark variables that contain an
autoCopy which should not be removed (i.e. a copy that is necessary for MM
correctness under asynchronous execution.
The FLAG_ATOMIC_TYPE was moved down, to restore local alphbeticization.

    compiler/main/checks.cpp
Added a check for consistency of the autoCopyMap.

    compiler/passes/parallel.cpp
Added "non-blocking on" functions to the conditional on line 191.  
The decision on whether to add the autoCopy/autoDestroy pair and the code to
actually perform the insertion was factored out into the new function
insertAutoCopyDestroyForTaskArg().  
The variable receiving the autoCopy is now marked with FLAG_NECESSARY_AUTO_COPY,
to prevent its later removal.  
The generated code was modified, so the result of the autoCopy call is now
assigned to the corresponding field in the arg bundle.  Previously, the result
was just ignored.
The code for inserting autoDestroy calls in the task function now relies on the
assumption that every function involved has only one caller.  No check was added
for this, but if the assumption fails, we'll start getting negative reference
counts or valgrind errors.
Removed the break statement on line 1108.  Because the locale arg (passed to an
"on" function) can now appear in an autoCopy call, we have to remove all
references to it, not just the first.

    compiler/optimizations/remoteValueForwarding.cpp
Added a filter to RVF is not applied to reference atomic types.  Atomic types
are replaced with references to same in resolveFormals(), so that updates to
atomics display the expected reference semantics.

    compiler/optimizations/removeUnnecessaryAutoCopyCalls.cpp
We do not remove an autoCopyCall if it has the FLAG_NECESSARY_AUTO_COPY flag.
I also turned the condition that rhs->argList.head != NULL into an assert.
The continue performs a slight optimization, but makes the logic more obscure,
so I removed it.

    test/users/vass/domain_with_on-all.chpl
    test/users/vass/domain_with_on-all.good
Added new test case provided by Vass.

[tested on users/vass/domain_with_on-all.chpl]
[tested on examples/ (std)]
[tested on / (no-local)]
[tested on domains/ distributions/ multilocal/ performance/ with CHPL_COMM=gasnet]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22438 3a8e244f-b0f2-452b-bcba-4c88e055c3ca